Some minor changes to the retopology process to manage particular cases (meshes with holes)

+++ src/fgt/edit_topo/edit_topodialog.h
@@ -127,7 +127,7 @@ class edit_topodialog : public QWidget
 		void updateFceTable(QList<Fce> list);
 
 		int getIterations();
-		double dist();
+		float dist();
 
 		bool isDEBUG();
